requirements.txt
streamlit
langchain
langchain-community
langchain-openai
python-dotenv
chromadb
sentence-transformers
pypdf
huggingface-hub
openai
tiktoken
PyMuPDF 